Fix some test results. Move some declarations. Add more utility functions.
