+++ computationalLib/lib/network_shared/dbCommon.cpp
@@ -158,117 +158,11 @@ ResType ExecuteSubTransaction(TransactionT &txn,
                                           e.what(), __FILE__, __LINE__), shared::Severity::info);
   }
 }
-
-Experiment::Experiment(pqxx::row &row) {
-  experiment_id = row["experiment_id"].as<IndexType>();
-  user_id = row["user_id"].as<IndexType>();
-
-  auto rr = row["status"].as<std::string>();
-  status = strToEnum(rr,kStrToExpStatus);
-  parameters = Json::Value(row["parameters"].as<std::string>());
-  created_at = *StrToTimepoint(row["created_at"].as<std::string>());
-  start_time = (!row["start_time"].is_null()) ?
-               std::optional<IndexType>{
-                   *StrToTimepoint(row["start_time"].as<std::string>())
-               } : std::nullopt;
-  end_time = (!row["end_time"].is_null()) ?
-               std::optional<IndexType>{
-                   *StrToTimepoint(row["end_time"].as<std::string>())
-               } : std::nullopt;
+std::string PaginateRequest(std::string_view initial_request, IndexType page_num, IndexType page_size) {
+  return fmt::format("{} LIMIT {} OFFSET {}",
+                     initial_request,
+                     page_size,
+                     (page_num - 1) * page_size);
 }
 
-Iteration::Iteration(pqxx::row &row) {
-
-  iteration_id = row["iteration_id"].as<IndexType>();
-  experiment_id = row["experiment_id"].as<IndexType>();
-  node_id = row["node_id"].as<IndexType>();
-
-  auto type=row["iter_t"].as<std::string>();
-  iter_t = strToEnum(type,kStrToIterType);
-
-  auto st=row["status"].as<std::string>();
-  status = strToEnum(st,kStrToIterStatus);
-
-  output_data = Json::Value(row["output_data"].as<std::string>());
-
-
-  start_time = *StrToTimepoint(row["start_time"].as<std::string>());
-
-  end_time = (!row["end_time"].is_null()) ?
-             std::optional<IndexType>{
-                 *StrToTimepoint(row["end_time"].as<std::string>())
-             } : std::nullopt;
-}
-
-
-User::User(pqxx::row &row) {
-  user_id = row["user_id"].as<IndexType>();
-  login = row["login"].as<std::string>();
-  hashed_password = row["hashed_password"].as<std::string>();
-  auto rr = row["role"].as<std::string>();
-  role = strToEnum(rr,kStrToUserRole);
-  user_id = row["user_id"].as<IndexType>();
-
-  created_at = *StrToTimepoint(row["created_at"].as<std::string>());
-  last_login = (!row["last_login"].is_null()) ?
-               std::optional<IndexType>{
-                   *StrToTimepoint(row["created_at"].as<std::string>())
-               } : std::nullopt;
-
-}
-
-bool User::operator==(const User &rhs) const {
-  return login == rhs.login &&
-      hashed_password == rhs.hashed_password &&
-      role == rhs.role;
-}
-
-bool Experiment::operator==(const Experiment &rhs) const {
-  return experiment_id == rhs.experiment_id;
-}
-
-bool Iteration::operator==(const Iteration &rhs) const {
-  return iteration_id==rhs.iteration_id;
-}
-
-bool Node::operator==(const Node &rhs) const {
-  return ip_address==rhs.ip_address;
-}
-Node::Node(pqxx::row &row) {
-
-  node_id = row["node_id"].as<IndexType>();
-  ip_address = row["ip_address"].as<std::string>();
-  auto sql_arr = row["benchmark_score"].as<std::string>();
-  auto arr = row["benchmark_score"].as_sql_array<shared::BenchResultType>();
-  using namespace print_utils;
-  benchmark_score =print_utils::ParseOneDimS<shared::BenchResVec>(sql_arr,
-                                                                  arr.size(),//todo use enum semntics
-                                                                  EIGENF(EigenPrintFormats::VECTOR_DB_FORMAT));
-  auto st=row["status"].as<std::string>();
-  status = strToEnum(st,kStrToNodeSt);
-
-  last_ping = (!row["last_ping"].is_null()) ?
-               std::optional<IndexType>{
-                   *StrToTimepoint(row["last_ping"].as<std::string>())
-               } : std::nullopt;
-}
-template<typename T>
-std::optional<T> getOpt(pqxx::field&&a)
-{
-  return (a.is_null())?std::nullopt:std::optional<T>{a.as<T>()};
-}
-Log::Log(pqxx::row &row) {
-  log_id=row["log_id"].as<IndexType>();
-  node_id=std::move(getOpt<IndexType>(row["node_id"]));
-  experiment_id=std::move(getOpt<IndexType>(row["experiment_id"]));
-
-  auto sev=row["severity"].as<std::string>();
-  severity = shared::Severity::info;//todo redo
-  message=row["message"].as<std::string>();
-
-  timestamp=*StrToTimepoint(row["timestamp"].as<std::string>());
-}
-bool Log::operator==(const Log &rhs) const {
-  return log_id==rhs.log_id;
-}
 }
