+++ utilLib/include/common/enumUtils.h
@@ -8,20 +8,52 @@
 
 namespace enum_utils {
 using shared::InvalidOption;
+
+/**
+ * @brief Structure to store enum mapping to some other class
+ * @tparam Enum
+ */
+template<typename Enum,typename To> requires std::is_enum_v<Enum>
+struct EnumDoubleMapping {
+  Enum value;
+  To to;
+};
 /**
  * @brief Structure to store enum mapping
  * @tparam Enum
  */
 template<typename Enum> requires std::is_enum_v<Enum>
-struct EnumMapping {
-  Enum value;
-  std::string str;
-};
-template<typename Enum> requires std::is_enum_v<Enum>
-using MapStrToEnum = std::unordered_map<std::string, Enum>;
+using EnumMapping=EnumDoubleMapping<Enum,std::string>;
 
-template<typename Enum> requires std::is_enum_v<Enum>
-using MapEnumToStr = std::unordered_map<Enum, std::string>;
+
+template<typename Enum,typename From> requires std::is_enum_v<Enum>
+using MapFromTypeToEnum = std::unordered_map<From, Enum>;
+
+template<typename Enum,typename To> requires std::is_enum_v<Enum>
+using MapEnumToType = std::unordered_map<Enum, To>;
+
+
+/**
+ * Generate map from From to Enum
+ * @tparam Enum
+ * @tparam From
+ * @param mappings
+ */
+template<typename Enum,typename From>
+requires std::is_enum_v<Enum>
+MapFromTypeToEnum<Enum, From>
+CreateFromTypeToEnumMap(const std::vector<EnumDoubleMapping<Enum,From>> &mappings);
+
+/**
+ * @brief Generate map from Enum to To type map
+ * @tparam Enum
+ * @tparam To
+ * @param mappings
+ */
+template<typename Enum,typename To>
+requires std::is_enum_v<Enum>
+MapEnumToType<Enum, To> CreateEnumToTypeMap(const std::vector<EnumDoubleMapping<Enum,To>> &mappings);
+
 
 /**
  * @brief Generate map from Str to enum value
@@ -30,7 +62,8 @@ using MapEnumToStr = std::unordered_map<Enum, std::string>;
  */
 template<typename Enum>
 requires std::is_enum_v<Enum>
-MapStrToEnum<Enum> createStrToEnumMap(const std::vector<EnumMapping<Enum>> &mappings);
+MapFromTypeToEnum<Enum, std::string>
+    CreateStrToEnumMap(const std::vector<EnumMapping<Enum>> &mappings);
 
 /**
  * Generate map from enum value to Str
@@ -40,18 +73,20 @@ MapStrToEnum<Enum> createStrToEnumMap(const std::vector<EnumMapping<Enum>> &mapp
  */
 template<typename Enum>
 requires std::is_enum_v<Enum>
-MapEnumToStr<Enum> createEnumToStrMap(const std::vector<EnumMapping<Enum>> &mappings);
+MapEnumToType<Enum, std::string> CreateEnumToStrMap(const std::vector<EnumMapping<Enum>> &mappings);
+
 
 /**
- * @brief P
+ * @brief Casts From type value to enum
  * @tparam Enum
+ * @tparam From
  * @param str
  * @param map
- * @return
+ * @return Enum value
  */
-template<typename Enum>
+template<typename Enum,typename From>
 requires std::is_enum_v<Enum>
-Enum inline strToEnum(const std::string &str, const std::unordered_map<std::string, Enum> &map) {
+Enum inline FromToEnum(const From str, const MapFromTypeToEnum<Enum,From> &map) {
   auto it = map.find(str);
   if (it == map.end()) {
     throw InvalidOption(str);
@@ -59,33 +94,84 @@ Enum inline strToEnum(const std::string &str, const std::unordered_map<std::stri
   return it->second;
 }
 
-template<typename Enum>
+/**
+ * @brief Casts Enum Value To typed value
+ * @tparam Enum
+ * @tparam To
+ * @param value
+ * @param map
+ * @return To value
+ */
+template<typename Enum,typename To>
 requires std::is_enum_v<Enum>
-std::string inline enumToStr(Enum value, const std::unordered_map<Enum, std::string> &map) {
+To inline EnumToType(Enum value, const std::unordered_map<Enum, To> &map) {
   auto it = map.find(value);
   return it->second;
+}
+
+/**
+ * @brief Casstst string to enum
+ * @tparam Enum
+ * @param str
+ * @param map
+ * @return Enum value
+ */
+template<typename Enum>
+requires std::is_enum_v<Enum>
+Enum inline StrToEnum(const std::string &str, const std::unordered_map<std::string, Enum> &map) {
+  return FromToEnum(str,map);
+}
+
+/**
+ * @brief Casts enum value to string
+ * @tparam Enum
+ * @param value
+ * @param map
+ * @return std::string
+ */
+template<typename Enum>
+requires std::is_enum_v<Enum>
+std::string inline EnumToStr(Enum value, const std::unordered_map<Enum, std::string> &map) {
+  return EnumToType(value,map);
 }
 
 }
 
 namespace enum_utils {
-template<typename Enum>
+
+
+template<typename Enum,typename From>
 requires std::is_enum_v<Enum>
-MapStrToEnum<Enum> createStrToEnumMap(const std::vector<EnumMapping<Enum>> &mappings) {
-  MapStrToEnum<Enum> map;
+MapFromTypeToEnum<Enum, From>
+CreateFromTypeToEnumMap(const std::vector<EnumDoubleMapping<Enum,From>> &mappings)
+{
+  MapFromTypeToEnum<Enum, From> map;
   for (const auto &mapping : mappings) {
-    map[mapping.str] = mapping.value;
+    map[mapping.to] = mapping.value;
+  }
+  return map;
+}
+
+template<typename Enum,typename To>
+requires std::is_enum_v<Enum>
+MapEnumToType<Enum, To> CreateEnumToTypeMap(const std::vector<EnumDoubleMapping<Enum,To>> &mappings)
+{
+  MapEnumToType<Enum, To> map;
+  for (const auto &mapping : mappings) {
+    map[mapping.value] = mapping.to;
   }
   return map;
 }
 
 template<typename Enum>
 requires std::is_enum_v<Enum>
-MapEnumToStr<Enum> createEnumToStrMap(const std::vector<EnumMapping<Enum>> &mappings) {
-  MapEnumToStr<Enum> map;
-  for (const auto &mapping : mappings) {
-    map[mapping.value] = mapping.str;
-  }
-  return map;
+MapFromTypeToEnum<Enum, std::string> CreateStrToEnumMap(const std::vector<EnumMapping<Enum>> &mappings) {
+  return CreateFromTypeToEnumMap(mappings);
+}
+
+template<typename Enum>
+requires std::is_enum_v<Enum>
+MapEnumToType<Enum, std::string> CreateEnumToStrMap(const std::vector<EnumMapping<Enum>> &mappings) {
+  return CreateEnumToTypeMap(mappings);
 }
 }
